cast region as uint16

+++ ibllib/atlas/regions.py
@@ -53,7 +53,7 @@ class BrainRegions(_BrainRegions):
                          rgb=c,
                          level=df_regions.depth.to_numpy(),
                          parent=df_regions.parent_structure_id.to_numpy(),
-                         order=df_regions.graph_order.to_numpy())
+                         order=df_regions.graph_order.to_numpy().astype(np.uint16))
         # mappings are indices not ids: they range from 0 to n regions -1
         mappings = pd.read_parquet(FILE_MAPPINGS)
         self.mappings = {k: mappings[k].to_numpy() for k in mappings}
